move file size calculation to string utils
