+++ src/utils/string/index.ts
@@ -59,3 +59,12 @@ export const processLogString = (logString: string) =>
  */
 export const stringIntersection = (string1: string, string2: string) =>
   string1.includes(string2) || string2.includes(string1);
+
+export const formatBytes = (bytes: number, decimals = 2) => {
+  if (bytes === 0) return "0 Bytes";
+  const k = 1024;
+  const dm = decimals < 0 ? 0 : decimals;
+  const sizes = ["Bytes", "KB", "MB", "GB"];
+  const i = Math.floor(Math.log(bytes) / Math.log(k));
+  return `${parseFloat((bytes / k ** i).toFixed(dm))} ${sizes[i]}`;
+};
